+++ src/apps/apps.cpp
@@ -18,8 +18,8 @@ void frij_register_apps(void)
     frij_registry_add(home_app());
     frij_registry_add(todo_app());
     frij_registry_add(events_app());
-    frij_registry_add(stopwatch_app());
     frij_registry_add(scoreboard_app());
+    frij_registry_add(stopwatch_app());
     frij_registry_add(assistant_app());
     frij_registry_add(counter_app());
     frij_registry_set_settings(settings_app());